+++ Exec/science/wdmerger/problem_tagging_nd.F90
@@ -46,9 +46,6 @@ contains
     integer  :: i, j, k, n
     real(rt) :: loc(3), r, r_P, r_S
 
-    logical  :: outer_boundary_test(3)
-    integer  :: boundary_buf(3), idx(3)
-
     !$gpu
 
     do k = lo(3), hi(3)
@@ -140,45 +137,6 @@ contains
              ! Clear all tagging that occurs outside the radius set by max_tagging_radius.
 
              if (r .gt. max_tagging_radius * max(maxval(abs(problo-center)), maxval(abs(probhi-center)))) then
-
-                tag(i,j,k) = clear
-
-             endif
-
-             ! We must ensure that the outermost zones are untagged
-             ! due to the Poisson equation boundary conditions.
-             ! We currently do not know how to fill the boundary conditions
-             ! for fine levels that touch the physical boundary.
-             ! (Note that this does not apply for interior/symmetry boundaries.)
-             ! To do this properly we need to be aware of AMReX's strategy
-             ! for tagging, which is not cell-based, but rather chunk-based.
-             ! The size of the chunk on the coarse grid is given by
-             ! blocking_factor / ref_ratio -- the idea here being that
-             ! blocking_factor is the smallest possible group of cells on a
-             ! given level, so the smallest chunk of cells possible on the
-             ! coarse grid is given by that number divided by the refinement ratio.
-             ! So we cannot tag anything within that distance from the boundary.
-             ! Additionally we need to stay a further amount n_error_buf away,
-             ! since n_error_buf zones are always added as padding around
-             ! tagged zones.
-
-             outer_boundary_test = .false.
-             boundary_buf(1:dim) = n_error_buf(level) + blocking_factor(level+1) / ref_ratio(1:dim, level)
-             idx = [i, j, k]
-
-             do n = 1, dim
-
-                if ((physbc_lo(n) .ne. Symmetry) .and. (idx(n) .le. domlo_level(n, level) + boundary_buf(n))) then
-                   outer_boundary_test(n) = .true.
-                endif
-
-                if ((physbc_hi(n) .ne. Symmetry) .and. (idx(n) .ge. domhi_level(n, level) - boundary_buf(n))) then
-                   outer_boundary_test(n) = .true.
-                endif
-
-             enddo
-
-             if ( any(outer_boundary_test) ) then
 
                 tag(i,j,k) = clear